+++ src/components/App/routes.js
@@ -6,6 +6,7 @@ import RankPage from "../RankPage/RankPage";
 import FolderEditPage from "../FolderEditPage/FolderEditPage";
 import Header from "../Header/Header";
 import Category from "./Category/Category";
+import Tree from "../Tree/Tree";
 
 const routes = (auth) => [
   {
@@ -24,7 +25,11 @@ const routes = (auth) => [
   {
     path: "/auth",
     element: <Outlet />,
-    children: [{ path: "/auth/login", element: <Category /> }],
+    children: [
+      { path: "/auth/login", element: <Category /> },
+      { path: "/auth/tree", element: <Tree /> },
+    ],
+
     // <Login auth={auth} />
   },
 ];
