+++ src/components/Tree/Tree.css
@@ -1,0 +1,56 @@
+.container {
+  display: flex;
+  width: 100%;
+  margin: 10px;
+}
+
+ul {
+  list-style: none;
+  border-radius: 25px;
+  background-color: rgb(255, 255, 255);
+  margin: 0;
+}
+
+li {
+  background-color: rgb(255, 255, 255);
+  border-radius: 15px;
+  padding: 10px;
+}
+
+.data-box,
+.history-box {
+  display: flex;
+  justify-content: center;
+  border-radius: 25px;
+  width: 45%;
+  margin: 0 auto;
+  padding: 15px;
+}
+
+.data-box {
+  max-height: 700px;
+  border-radius: 25px;
+  background-color: #5587f5;
+}
+
+.history-box {
+  overflow: hidden;
+  border-radius: 25px;
+  background-color: #f2c84d;
+}
+
+.folder-list {
+  display: flex;
+  justify-content: center;
+  width: 80%;
+}
+
+.folder {
+  width: 100%;
+  height: 30px;
+}
+
+.drag-target {
+  background-color: rgb(184, 184, 250);
+  cursor: grabbing;
+}
